--- src/components/home-page/home-page.tsx
@@ -1,7 +1,0 @@
-export interface HomePageProps {
-    className?: string;
-}
-
-export const HomePage = ({ className }: HomePageProps) => {
-    return <div className={className}>Home</div>;
-};
