+++ src/components/notes/fake-note/fake-note.module.scss
@@ -1,0 +1,19 @@
+@import '../../../styles/common/variables.module.scss';
+
+.root {
+    display: flex;
+    flex-direction: column;
+    flex-wrap: nowrap;
+    gap: $size4;
+    width: 200px;
+    height: 180px;
+    background-color: $sand1;
+    border-radius: 4px;
+    border: 1px solid $grass1;
+    align-items: center;
+    justify-content: center;
+}
+
+.root:hover {
+    border: 2px solid $grass4;
+}
